+++ gen_class.py
@@ -4,6 +4,9 @@
 
     Description:  Class that has class definitions for general use.
 
+    Function:
+        setup_mail
+
     Classes:
         Daemon
         ProgressBar
@@ -30,6 +33,8 @@ import time
 import atexit
 import signal
 import platform
+import socket
+import getpass
 
 # Third-party
 import yum
@@ -38,6 +43,33 @@ import yum
 import version
 
 __version__ = version.__version__
+
+
+def setup_mail(to_line, subj=None, frm_line=None, **kwargs):
+
+    """Function:  setup_mail
+
+    Description:  Initialize a mail instance.  Provide 'from line' if one is
+        not passed.
+
+    Arguments:
+        (input) to_line -> Mail to line.  Either a string or list.
+        (input) subj -> Mail subject line.  Either a string or list.
+        (input) frm_line -> Mail from line.
+        (output) Mail instance.
+
+    """
+
+    if isinstance(to_line, list):
+        to_line = list(to_line)
+
+    if isinstance(subj, list):
+        subj = list(subj)
+
+    if not frm_line:
+        frm_line = getpass.getuser() + "@" + socket.gethostname()
+
+    return Mail(to_line, subj, frm_line)
 
 
 class Daemon: